Revert "fix(394): prevent mocked query results from returning undefined"
This reverts commit b6c8bd8ea30006cceedec742447e5605a34d08e4.
